requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.29.7
psycopg[binary]==3.2.12
python-dotenv==1.0.0
openai==1.3.7
pytest==7.4.3
pytest-cov==4.1.0

